TM-3758 Workflow step - Double click on delete action throw an exception page
git-svn-id: https://opsvcs.tdsops.com/repos/tdstm/trunk@5674 fecba46a-3362-0410-80e5-8c76cdb905a0

+++ grails-app/controllers/WorkflowController.groovy
@@ -338,15 +338,17 @@ def updateWorkflowSteps() {
 		def transitionId = params.id
 		if(transitionId){
 			def workflowTransition = WorkflowTransition.get( transitionId )
-			def process = workflowTransition.workflow.process
-			//AssetTransition.executeUpdate('delete from')
-			StepSnapshot.executeUpdate("delete from StepSnapshot ss where ss.moveBundleStep in (select mbs.id from MoveBundleStep mbs where mbs.moveBundle.project.workflowCode = '${workflowTransition.workflow.process}' and mbs.transitionId = ${workflowTransition.transId})")
-			MoveBundleStep.executeUpdate("delete from MoveBundleStep mbs where mbs.moveBundle in (select mb.id from MoveBundle mb where mb.workflowCode = '${workflowTransition.workflow.process}') and mbs.transitionId = ?",[ workflowTransition.transId])
-			WorkflowTransitionMap.executeUpdate("delete from WorkflowTransitionMap wtm where wtm.workflowTransition = ?",[workflowTransition])
-			workflowTransition.delete(flush:true)
-			
-			//	load transitions details into application memory.
-	    	stateEngineService.loadWorkflowTransitionsIntoMap( process, 'workflow')
+			if (workflowTransition) {
+				def process = workflowTransition.workflow.process
+				//AssetTransition.executeUpdate('delete from')
+				StepSnapshot.executeUpdate("delete from StepSnapshot ss where ss.moveBundleStep in (select mbs.id from MoveBundleStep mbs where mbs.moveBundle.project.workflowCode = '${workflowTransition.workflow.process}' and mbs.transitionId = ${workflowTransition.transId})")
+				MoveBundleStep.executeUpdate("delete from MoveBundleStep mbs where mbs.moveBundle in (select mb.id from MoveBundle mb where mb.workflowCode = '${workflowTransition.workflow.process}') and mbs.transitionId = ?",[ workflowTransition.transId])
+				WorkflowTransitionMap.executeUpdate("delete from WorkflowTransitionMap wtm where wtm.workflowTransition = ?",[workflowTransition])
+				workflowTransition.delete(flush:true)
+				
+				//	load transitions details into application memory.
+		    	stateEngineService.loadWorkflowTransitionsIntoMap( process, 'workflow')
+		    }
 		}
 		redirect(action:"workflowList", params:[workflow:workflowId])
 	}
